Telegram Bot to receive commands and words
The commands are then commited to the `telegram-command-trigger` topic.
If a user doesn't send a /kafka command, but free text, the input text
is then committed to the `word-count`-Stream. From there it can be
easily read by the WordCounter Applicattion, which processes the input
and sends the result to the `word-count-result` topic.